extract sidebar gui into own class
